Disable windows build as it does not work and tries to dl its own openssl
The configure.win script tries to download it's own precompiled openssl
libs from the internet and fails. Disabling as this 1. does not work and
2. is not a good practice.

The previous version does also not compile anymore for the same reason. So
not updating is also not an option.

Using the conda openssl also on windows seems not to work out of the box
with the current configure.win script.